requirements.txt
pymupdf
rapidfuzz
rank-bm25
pydantic[email]
argon2-cffi
python-multipart
langchain-text-splitters
cryptography
deep-translator
fastapi
firebase-admin
httpx
langchain==0.1.20 
langchain-community==0.0.38 
langchain-core==0.1.52
langchain-google-genai
faiss-cpu
pandas
PyJWT
python-dotenv
bandit
safety
pip-audit
detect-secrets
uvicorn
google-cloud-secret-manager
google-cloud-firestore
google-cloud-storage
gunicorn